custom serialzation: - improve some comments - adjust some code - move the rpc call stuff - include call_async
